+++ readium/shared/src/main/java/org/readium/r2/shared/util/archive/channel/ChannelZipContainer.kt
@@ -130,7 +130,7 @@ internal class ChannelZipContainer(
             .filterNot { it.isDirectory }
             .mapNotNull { Entry(it) }
 
-    override suspend fun get(path: String): Container.Entry =
+    override fun get(path: String): Container.Entry =
         archive.getEntry(path.removePrefix("/"))
             ?.takeUnless { it.isDirectory }
             ?.let { Entry(it) }